make parsing ignore sequence digits
Ref SPIN-1754. For some use-cases we may need
to support a larger number of digits for group
sequence numbers. This change makes the parsing
more lenient to the number of digits in the
sequence.

Tools such as spinnaker that are creating the
resources would be responsible for figuring out
the appropriate number of digits and worrying
about padding. This change is to allow tools
that just consume the group names to move forward.