Address https://github.com/spinnaker/spinnaker/issues/602 by making frigga configurable
The idea is to define in an external configuration the following properties :
*  `DELIMITER`  delimiter used in infrastructure items names (ASG, ELB, ...) : currently hard coded to `-` in Frigga
*  `APPVERSION_DELIMITER`  delimiter used in app versions: currently hard coded to `-` in Frigga
*  `AUTHORIZED_CHARS`  authorized chars in infrastructure items names (ASG, ELB, ...) : currently hard coded to `._` in Frigga
*  `APPVERSION_AUTHORIZED_CHARS`  authorized chars in app versions : currently hard coded to `._` in Frigga

I duplicated the existing tests `NamesSpec` and `NameValidationSpec` , swap `-` by `_` in tests data, then replace the default values in `NameConstants` for POCing my idea with :
`DELIMITER` = `_`  and `AUTHORIZED_CHARS` = `.\\-` . The new tests went green.

Currently I tried to use Env vars to set the configurable properties but it would be better to use a config file and add tests with different configurations.