add support for ~ in detail
Follow up to #18 and Netflix/atlas#513 adding support
for the `~` character as part of the detail. The general
agreement is there will be no further relaxation of the
rules for group names.

This change also updates the test cases and name pattern
to fix a bug where using a `~` or `^` in the stack was
causing it to be treated as the detail instead of failing
to match. Before:

```
application-^stack-detail

- app = "application"
- stack = null
- detail = "^stack-detail"
```

Now it will be treated just like any other invalid group
name.