refactor(labeledVariables): extract labeled variables to a NamingConvention
Breaking

Accessor methods still exist on Names and calling one will parse labeled variables out of the cluster name, preserving
the existing values as they would be returned.

Labeled variables are not dropped prior to considering the app/stack/detail components of a name, so where previously a
cluster called foo-c0banana would have a stack of null and a country of banana, now will have a stack of c0banana and
a country of banana.